tests/wpt/tests/html/canvas/offscreen/fill-and-stroke-styles/2d.gradient.object.return.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.gradient.object.return</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.gradient.object.return</h1>
<p class="desc">createLinearGradient() and createRadialGradient() returns objects implementing CanvasGradient</p>


<script>
var t = async_test("createLinearGradient() and createRadialGradient() returns objects implementing CanvasGradient");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  window.CanvasGradient.prototype.thisImplementsCanvasGradient = true;

  var g1 = ctx.createLinearGradient(0, 0, 100, 0);
  _assertDifferent(g1.addColorStop, undefined, "g1.addColorStop", "undefined");
  _assertSame(g1.thisImplementsCanvasGradient, true, "g1.thisImplementsCanvasGradient", "true");

  var g2 = ctx.createRadialGradient(0, 0, 10, 0, 0, 20);
  _assertDifferent(g2.addColorStop, undefined, "g2.addColorStop", "undefined");
  _assertSame(g2.thisImplementsCanvasGradient, true, "g2.thisImplementsCanvasGradient", "true");
  t.done();

});
</script>
